web-platform_22_02_dev/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
comtypes==1.1.10
cycler==0.11.0
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
fastkml==0.12
fonttools==4.28.3
frozenlist==1.2.0
future==0.18.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.13.3
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.5.0
multidict==5.2.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pygeoif==0.7
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==3.0.6
PyPDF2==1.26.0
pypiwin32==223
PySide6==6.2.2
python-bidi==0.4.2
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
reportlab==3.6.3
requests==2.26.0
scikit-image==0.19.0
scipy==1.7.3
setuptools-scm==6.3.2
shiboken6==6.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tifffile==2021.11.2
tomli==1.2.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
yarl==1.7.2